// web/src/lib/api/models/app/emails/MailboxImport.ts
export type MailHost =
    | ""
    | "google_workspace"
    | "gmail"
    | "microsoft365"
    | "outlook"
    | "zoho"
    | "yahoo"
    | "aol"
    | "icloud"
    | "fastmail"
    | "godaddy"
    | "namecheap"
    | "ionos"
    | "hostinger"
    | "ovh"
    | "migadu"
    | "purelymail"
    | "rackspace"
    | "yandex"
    | "gmx"
    | "proton"
    | "other";

export type MailAuthMethod = "" | "password" | "app_password" | "oauth" | "delegated";